refactor: PhonemeServer
